requirements.txt
openai>=1.50.0
anthropic>=0.30.0
google-genai>=1.0.0
PyGithub>=2.0.0
